• I added a game object class to contain the graphics and physics part of the objects. • I added a manager to manage materials.

+++ Assignment2/EGP300-proj1/GUIImage.cpp
@@ -38,7 +38,7 @@ void GUIImage::Initialize(Sprite2D* image, float x, float y)
 {
 	Vector2f position = Vector2f(x,y);
 	mTransform = new Transform2D(position);
-	mTransform->SetScale(image->getWidth(), image->getHeight());
+	mTransform->SetScale((float)image->getWidth(), (float)image->getHeight());
 	mVisible = true;
 	mSprite = image;
 }
